+++ src/common/pa_cpuload.h
@@ -46,7 +46,8 @@
  stream. Used to implement the Pa_GetStreamCpuLoad() function.
 */
 
-
+#include <stdint.h>
+#include <stddef.h>
 #ifdef __cplusplus
 extern "C"
 {
@@ -61,7 +62,7 @@ typedef struct {
 
 void PaUtil_InitializeCpuLoadMeasurer( PaUtilCpuLoadMeasurer* measurer, double sampleRate );
 void PaUtil_BeginCpuLoadMeasurement( PaUtilCpuLoadMeasurer* measurer );
-void PaUtil_EndCpuLoadMeasurement( PaUtilCpuLoadMeasurer* measurer, unsigned long framesProcessed );
+void PaUtil_EndCpuLoadMeasurement( PaUtilCpuLoadMeasurer* measurer, size_t framesProcessed );
 void PaUtil_ResetCpuLoadMeasurer( PaUtilCpuLoadMeasurer* measurer );
 double PaUtil_GetCpuLoad( PaUtilCpuLoadMeasurer* measurer );
