+++ src/os/unix/pa_unix_util.c
@@ -71,7 +71,7 @@ static int numAllocations_ = 0;
 #endif
 
 
-void *PaUtil_AllocateMemory( long size )
+void *PaUtil_AllocateMemory( size_t  size )
 {
     void *result = malloc( size );
